Created Vehicular Chassis to Constrain Wheels
Hi Abraham and Jason, thought of a quick fix to constrain the bodies. I built a chassis "wall" to cage the spherical wheels and let physics engine do the rest. You may have to manually put in the ground plane again to try the scene as I am trying to reduce the number of files to upload. We can discuss more later.